Better locale axis time formatting (#241)
* Initial changes to support custom formats for absolute time

* Code cleanup and handling custom relative time formatting.

* Introducing a configurable time epoch.

* Docs and tests (and minor fixes).

* Have to make everything lowercase since Chrome and Firefox case 'dayPeriod' differently.

* Change how time epoch is specified. Remove "fix" for centering multi-line text (it had bugs).

* Fixing linting errors.

* Removing extraneous formatString option. Adding relative time tests.

* Removing old mechanism for formatting time.

Globalize.format would format a date offset by the unix epoch with a year of "0000". This is silly (and wrong). The new mechanism will properly give a year of '0001'.

* Putting back the old formatting code.

It's too much of a pain to deal with this right now.  All of the old tests are written against "year 0", and thus are expecting a silly result. I'll fix this later.

* Documentation change. Getting rid of unnecessary variable.

* Missed getting rid of space at the end.