Removed global options object in flot.legend
This fixes an issue when using multiple plots with legends. Before this change, the global options object would refer to the options of the last plot that called init. Because of this, when multiple plots would execute the setupGrid hook, it could insert the legend elements into the wrong container.